pyproject.toml
[tool.poetry]
name = "ga-prospective-study"
version = "0.1.0"
description = ""
authors = ["bbearce <bbearce@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
bpython = "^0.24"
seaborn = "^0.13.2"
matplotlib = "^3.9.1"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.1"
pingouin = "^0.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
